intiial refactoring of styles
