+++ src/PageRouter.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react';
 import { useContext } from './contexts';
 import { FiZapOff, FiZap } from 'react-icons/fi';
 import { useLocation, useHistory } from 'react-router-dom';
-import Box from './components/common/Box';
+import { Box } from './components/common';
 
 const PageRouterWrapper = styled.div`
   margin-top: 20px;
