+++ src/components/Header/index.tsx
@@ -7,7 +7,7 @@ import { FiSettings, FiUser, FiBarChart2 } from 'react-icons/fi';
 import dxdaoIcon from 'assets/images/DXdao.svg';
 import Web3 from 'web3';
 import { bnum } from '../../utils';
-import Box from '../../components/common/Box';
+import { Box } from '../../components/common';
 
 const NavWrapper = styled.div`
   display: flex;
