+++ src/components/common/InfoBox.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const InfoBox = styled.div`
+  flex: 1;
+  text-align: center;
+  padding: 2px 5px;
+  margin: 10px 5px;
+  font-size: 25px;
+  font-weight: 300;
+  border-radius: 3px;
+  color: var(--activeButtonBackground);
+`;
