+++ src/components/common/Row.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Row = styled.div`
+  flex-direction: row;
+  display: flex;
+  padding-top: 15px;
+  justify-content: space-around;
+`;
