+++ src/components/common/index.ts
@@ -1,0 +1,6 @@
+export { Row } from './Row';
+export { Box } from './Box';
+export { InfoBox } from './InfoBox';
+export { Question } from './Question'
+export { BlockchainLink } from './BlockchainLink';
+export { ActiveButton } from './ActiveButton';
