+++ src/config.json
@@ -2456,22 +2456,22 @@
   "localhost": {
     "cache": {
       "fromBlock": 0,
-      "toBlock": 2,
-      "ipfsHash": "QmdDK53AiMxugzWXKziZBw5CCnHunWc9Z41RiaE8Z4XCGL"
+      "toBlock": 68,
+      "ipfsHash": "QmbpBYrD3CGyxZPxWA2wBc1aJfDi16Cvjb2KBD5xrwKXHu"
     },
     "contracts": {
-      "avatar": "0x1C469EDDcd7d058be7195B2b321675bd1521AECf",
-      "controller": "0xD98034E78C052C9B975E3D1C3d274e4ab54A8092",
-      "reputation": "0x70cC43912fAD6842E9dE9dfb1e8AaF92EDbBc40C",
+      "avatar": "0xB5AE3e2883934457cb593793f034b401867e1372",
+      "controller": "0x56F7DC1Cc938d6d6575B68d70d381F9C28C8C7b7",
+      "reputation": "0x50D75C1BC6a1cE35002C9f92D0AF4B3684aa6B74",
       "votingMachines": {
         "dxd": {
-          "address": "0x742121cD0BB9784c1f17f25B8da1bF3445b93ED4",
-          "token": "0x64E15dFA18DDE38F4D4e27EFc00F4BCFCCe8144a"
+          "address": "0xD63259B15E88DFDC994A90104a2c7226CE3ECa69",
+          "token": "0x668eD30aAcC7C7c206aAF1327d733226416233E2"
         }
       },
-      "permissionRegistry": "0x845856776D110a200Cf41f35C9428C938e72E604",
+      "permissionRegistry": "0x69E40e93f68c336B97B6C0583C33162DB3416618",
       "utils": {
-        "multicall": "0xb5577DD27eF3690f1fE0f5e3a46125f73aCe8Afa"
+        "multicall": "0x13E07a1dC038B61cbDb1D501221Ec634cC4a7013"
       }
     },
     "recommendedCalls": [],
