+++ src/pages/FAQ.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import styled from 'styled-components';
 import { observer } from 'mobx-react';
-import Box from '../components/common/Box';
+import { Box } from '../components/common';
 import { useContext } from '../contexts';
 import { useLocation } from 'react-router-dom';
 import MDEditor from '@uiw/react-md-editor';
