+++ src/pages/NewProposal.tsx
@@ -2,9 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { observer } from 'mobx-react';
 import { useContext } from '../contexts';
-import ActiveButton from '../components/common/ActiveButton';
-import Question from '../components/common/Question';
-import Box from '../components/common/Box';
+import {Box, Question, ActiveButton } from '../components/common';
 import MDEditor, { commands } from '@uiw/react-md-editor';
 import contentHash from 'content-hash';
 import { NETWORK_ASSET_SYMBOL } from '../utils';
